cleaning up MappedSetting
