+++ MatterControlLib/PrinterCommunication/Io/PauseHandlingStream.cs
@@ -258,7 +258,7 @@ namespace MatterHackers.MatterControl.PrinterCommunication.Io
 
 		private void InjectPauseGCode(string codeToInject)
 		{
-			codeToInject = GCodeProcessing.ReplaceMacroValues(codeToInject);
+			codeToInject = printer.ReplaceMacroValues(codeToInject);
 
 			codeToInject = codeToInject.Replace("\\n", "\n");
 			string[] lines = codeToInject.Split('\n');
