+++ MatterControlLib/SlicerConfiguration/MappingClasses/ExtruderOffsets.cs
@@ -1,0 +1,84 @@
+﻿/*
+Copyright (c) 2016, Lars Brubaker
+All rights reserved.
+
+Redistribution and use in source and binary forms, with or without
+modification, are permitted provided that the following conditions are met:
+
+1. Redistributions of source code must retain the above copyright notice, this
+   list of conditions and the following disclaimer.
+2. Redistributions in binary form must reproduce the above copyright notice,
+   this list of conditions and the following disclaimer in the documentation
+   and/or other materials provided with the distribution.
+
+THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
+ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
+WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
+DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR
+ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
+(INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
+LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND
+ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+(INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
+SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+
+The views and conclusions contained in the software and documentation are those
+of the authors and should not be interpreted as representing official policies,
+either expressed or implied, of the FreeBSD Project.
+*/
+
+using System.Text;
+
+namespace MatterHackers.MatterControl.SlicerConfiguration.MappingClasses
+{
+	public class ExtruderOffsets : MappedSetting
+	{
+		public ExtruderOffsets(PrinterConfig printer, string canonicalSettingsName, string exportedName)
+			: base(printer, canonicalSettingsName, exportedName)
+		{
+		}
+
+		public override string Value
+		{
+			get
+			{
+				// map from 0x0,0x0,0x0
+				// to [[0,0],[0,0]]
+				StringBuilder final = new StringBuilder("[");
+				string[] offsets = base.Value.Split(',');
+				bool first = true;
+				int count = 0;
+				foreach (string offset in offsets)
+				{
+					if (!first)
+					{
+						final.Append(",");
+					}
+					string[] xy = offset.Split('x');
+					if (xy.Length == 2)
+					{
+						double x = 0;
+						double.TryParse(xy[0], out x);
+						double y = 0;
+						double.TryParse(xy[1], out y);
+						final.Append($"[{x},{y}]");
+						first = false;
+						count++;
+					}
+					else
+					{
+						final.Append("[0,0]");
+					}
+				}
+				while (count < 16)
+				{
+					final.Append(",[0,0]");
+					count++;
+				}
+				final.Append("]");
+
+				return final.ToString();
+			}
+		}
+	}
+}
