+++ MatterControlLib/SlicerConfiguration/MappingClasses/SupportExtrusionWidth.cs
@@ -1,0 +1,68 @@
+﻿/*
+Copyright (c) 2016, Lars Brubaker
+All rights reserved.
+
+Redistribution and use in source and binary forms, with or without
+modification, are permitted provided that the following conditions are met:
+
+1. Redistributions of source code must retain the above copyright notice, this
+   list of conditions and the following disclaimer.
+2. Redistributions in binary form must reproduce the above copyright notice,
+   this list of conditions and the following disclaimer in the documentation
+   and/or other materials provided with the distribution.
+
+THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
+ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
+WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
+DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR
+ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
+(INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
+LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND
+ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+(INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
+SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+
+The views and conclusions contained in the software and documentation are those
+of the authors and should not be interpreted as representing official policies,
+either expressed or implied, of the FreeBSD Project.
+*/
+
+namespace MatterHackers.MatterControl.SlicerConfiguration.MappingClasses
+{
+	public class SupportExtrusionWidth : MappedSetting
+	{
+		public SupportExtrusionWidth(PrinterConfig printer, string canonicalSettingsName, string exportedName)
+			: base(printer, canonicalSettingsName, exportedName)
+		{
+		}
+
+		public override string Value
+		{
+			get
+			{
+				double nozzleDiameter = printer.Settings.GetValue<double>(SettingsKey.nozzle_diameter);
+
+				string extrusionWidth = base.Value;
+
+				if (extrusionWidth == "0")
+				{
+					return "100";
+				}
+
+				if (extrusionWidth.Contains("%"))
+				{
+					string withoutPercent = extrusionWidth.Replace("%", "");
+					return withoutPercent;
+				}
+
+				double originalValue;
+				if (!double.TryParse(extrusionWidth, out originalValue))
+				{
+					originalValue = nozzleDiameter;
+				}
+
+				return (originalValue / nozzleDiameter * 100).ToString();
+			}
+		}
+	}
+}
